requirements.txt
# Core dependencies
dnspython>=2.6.0
openpyxl>=3.1.0

# Testing dependencies
pytest>=7.4.0
hypothesis>=6.92.0
pytest-mock>=3.12.0

# Build dependencies
pyinstaller>=6.3.0
